requirements.txt
# Core dependencies
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0

# Voice assistant stack – keep Pi-friendly versions
pvporcupine==1.9.0
pyaudio==0.2.11
openai-whisper==20230314
pyttsx3==2.90

# Optional (comment out to slim install)
mpg123==0.0.0  # placeholder for python wrapper if needed

# Utilities & development (can be removed for prod)
colorama==0.4.6
tqdm==4.64.0
